@@ -13,7 +13,7 @@ keywords = ["Ocean", "Particles", "Imaging", "Measurement", "Size distribution"]
 packages = [{include = "pyopia"}]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.11"
+python = "^3.12"
 flake8 = "^6.1.0"
 numpy = "^1.24.0"
 scipy = "^1.11.2"
@@ -42,16 +42,15 @@ typer = {extras = ["all"], version = "^0.9.0"}
 pandas = {version = "^2.1.1", extras = ["computation"]}
 h5py = "^3.9.0"
 poetry-version-plugin = "^0.2.0"
-tensorflow-macos = {version = "2.11.0", optional = true, markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
-tensorflow-cpu = {version = "2.11.0", optional = true}
-tensorflow-io-gcs-filesystem = [
-    {version = ">=0.31.0", optional=true}
-]
+tensorflow-macos = {version = "^2.16.2", optional = true, markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
+tensorflow-cpu = {version = "^2.16.2", optional = true}
 dask = ">=2024.8.1"
+nbconvert = "^7.16.4"
+h5netcdf = ">= 1.3.0"
 
 [tool.poetry.extras]
-classification-arm64 = ["tensorflow-io-gcs-filesystem", "tensorflow-macos"]
-classification = ["tensorflow-io-gcs-filesystem", "tensorflow-cpu"]
+classification-arm64 = ["tensorflow-macos"]
+classification = ["tensorflow-cpu"]
 
 [tool.poetry-version-plugin]
 source = "init"
