@@ -6,6 +6,7 @@ A high level test for the basic processing pipeline.
 from glob import glob
 import tempfile
 import os
+from tqdm import tqdm
 
 import pyopia.exampledata as exampledata
 import pyopia.io
@@ -18,6 +19,9 @@ import pandas as pd
 import skimage.io
 import numpy as np
 import pyopia.instrument.silcam
+
+
+ACCURACY = 39
 
 
 def test_match_to_database():
@@ -38,7 +42,7 @@ def test_match_to_database():
         model_path = exampledata.get_example_model(os.path.join(tempdir, 'model'))
 
         # Load the trained tensorflow model and class names
-        cl = pyopia.classify.Classify(model_path)
+        cl = pyopia.classify.Classify(model_path=model_path)
         class_labels = cl.class_labels
 
         # class_labels should match the training data
@@ -62,8 +66,9 @@ def test_match_to_database():
             t1 = pd.Timestamp.now()
 
             # loop through the database images
-            for file in files:
+            for file in tqdm(files):
                 img = skimage.io.imread(file)  # load ROI
+                img = np.float64(img) / 255
                 prediction = cl.proc_predict(img)  # run prediction from silcam_classify
 
                 ind = np.argmax(prediction)  # find the highest score
@@ -87,51 +92,60 @@ def test_match_to_database():
             name = os.path.split(cat)[-1]
             success = correct_positives(name)
             print(name, success)
-            assert success > 96, (name + ' was poorly classified at only ' + str(success) + 'percent.')
+            assert success > ACCURACY, (name + ' was poorly classified at only ' + str(success) + 'percent.')
 
 
 def test_pipeline_classification():
+    '''Check that the pipeline doesn't change the outcome of the classification.
+    Do this by putting rois of know classificion (which we know get correctly classified independintly),
+    and then use the same model in a pipeline analysing the synthetic image.
+    '''
+
     with tempfile.TemporaryDirectory() as tempdir:
         # location of the training data
         database_path = os.path.join(tempdir, 'silcam_classification_database')
 
         exampledata.get_classifier_database_from_pysilcam_blob(database_path)
-        os.makedirs(os.path.join(tempdir, 'model'), exist_ok=True)
-        model_path = exampledata.get_example_model(os.path.join(tempdir, 'model'))
+        os.makedirs('model', exist_ok=True)
+        model_path = exampledata.get_example_model('model')
 
         # Load the trained tensorflow model and class names
-        cl = pyopia.classify.Classify(model_path)
+        cl = pyopia.classify.Classify(model_path=model_path)
 
         def get_good_roi(category):
             '''
             calculate the percentage positive matches for a given category
             '''
 
+            print('category', category)
             # list the files in this category of the training data
             files = sorted(glob(os.path.join(database_path, category, '*.tiff')))
+            print(len(files), 'files')
 
+            found_match = 0
             # loop through the database images
-            for file in files:
-                img = skimage.io.imread(file)  # load ROI
+            for file in tqdm(files):
+                img = np.uint8(skimage.io.imread(file))  # load ROI
+                img = np.float64(img) / 255
                 prediction = cl.proc_predict(img)  # run prediction from silcam_classify
 
-                if np.max(prediction) < 0.96:
+                if np.max(prediction) < (ACCURACY / 100):
                     continue
 
                 ind = np.argmax(prediction)  # find the highest score
 
                 # check if the highest score matches the correct category
                 if cl.class_labels[ind] == category:
+                    print('roi file', file)
                     return img, category
+            assert found_match == 1, f'classifier not finding matching particle for {category}'
 
-        cl = pyopia.classify.Classify(model_path)
         canvas = np.ones((2048, 2448, 3), np.float64)
 
         rc_shift = int(2048/len(cl.class_labels)/1.5)
         rc = rc_shift
 
         classes = sorted(glob(os.path.join(database_path, '*')))
-        print(classes)
 
         categories = []
 
@@ -141,7 +155,7 @@ def test_pipeline_classification():
             categories.append(category)
             img_shape = np.shape(img)
             rc += rc_shift
-            canvas[rc:rc + img_shape[0], rc: rc + img_shape[1], :] = np.float64(img)/255
+            canvas[rc:rc + img_shape[0], rc: rc + img_shape[1], :] = np.float64(img)
 
         settings = {'general': {'raw_files': None,
                                 'pixel_size': 24},
@@ -167,6 +181,7 @@ def test_pipeline_classification():
 
         MyPipeline.data['imraw'] = canvas
         MyPipeline.data['timestamp'] = pd.Timestamp.now()
+        MyPipeline.data['filename'] = ''
 
         # Add the imageprep step description
         MyPipeline.settings['steps'].update({'imageprep':
@@ -181,7 +196,8 @@ def test_pipeline_classification():
         # Add the segmentation step description
         MyPipeline.settings['steps'].update({'segmentation':
                                             {'pipeline_class': 'pyopia.process.Segment',
-                                             'threshold': 1}})
+                                             'threshold': 1,
+                                             'segment_source': 'im_minimum'}})
         # Run the step
         MyPipeline.run_step('segmentation')
         # This is the same as running:
@@ -190,7 +206,8 @@ def test_pipeline_classification():
 
         # Add the segmentation step description
         MyPipeline.settings['steps'].update({'statextract':
-                                            {'pipeline_class': 'pyopia.process.CalculateStats'}}
+                                            {'pipeline_class': 'pyopia.process.CalculateStats',
+                                             'roi_source': 'imref'}}
                                             )
 
         # Run the step
@@ -202,9 +219,11 @@ def test_pipeline_classification():
         stats = pyopia.statistics.add_best_guesses_to_stats(MyPipeline.data['stats'])
 
         out = [x[12:] for x in stats['best guess'].values]
-        assert categories == out, 'Classes returned from classifier do not match what was given to the pipeline'
+
         print('classes input', categories)
         print('classes measured', out)
+
+        assert categories == out, 'Classes returned from classifier do not match what was given to the pipeline'
 
 
 if __name__ == "__main__":
